WIP refactored resizable widget
